Fix: Crash for user flow from Main menu to Credits and then Skip button
Edge case that had not been considered for user entering credits from main menu and then hitting the skip button in the credits scene. This would then try to go to the next level which is incorrect and would crash. The fix is a check for the previous scene to determine if coming from the main menu then just directly transition back to the main menu when skipping