Fix: Crash when opening door to courtyard in level 1
The door InteractableObject had empty transition objects set cause null reference errors. Guessing at some point there was a transition before the courtyard section was added and this was not changed. Removed the empty slots in the array transition so that no transition happens as its not needed here. Removed much of the debug stuff as this was outputting thousands of entries making it difficult to debug