Fix: Fixed jumping animation
Noticed the jump animation wasn't playing, so fixed the code to allow it to work by not setting to Idle if not moving left or right - as it could be jumping or trying to! Adjusted the animation so that it ends before touching ground so it doesn't look odd. Still has a bit of jerk at the bottom due sprite misalignment for jumps